+++ src/contrib/mod.rs
@@ -1,2 +1,5 @@
 #[cfg(feature = "amiwo_rocket")]
 pub mod rocket;
+
+#[cfg(feature = "amiwo_hyper")]
+pub mod hyper;