+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Version 0.2.0 ()
+- Added `GenericError` and changed all method signature to return a `GenericError`
+- Added `Pushable` trait and associated implementations for `serde_json::Value`
+- Refactored `FormHashMap` using 'serde_json::Value' instead of HashMap => now returns 'Value' instead of 'OneOrMany' types. FormHashMap can now parse 'application/json' data
+- Added utility function to `contrib/hyper` to create request and parse Response into ResponseJSON
+
 # Version 0.1.0 (May 19, 2017)
 - Created `amiwo::contrib` to hold contribution to other crates (currently rocket); moved `FormHashMap` to that module
 - Added `contrib::rocket::ResponseJSON` that can be used as a parameter and response with Rocket